refactor moshi classes/adapters
