--- app/src/main/java/com/example/android/temptext/network/ConditionDetails.kt
@@ -1,7 +1,0 @@
-package com.example.android.temptext.network
-
-import com.squareup.moshi.Json
-
-class ConditionDetails (
-
-)
