--- app/src/main/java/com/example/android/temptext/network/WeatherConditions.kt
@@ -1,8 +1,0 @@
-package com.example.android.temptext.network
-
-import com.squareup.moshi.Json
-
-class WeatherConditions (
-    @Json(name = "text") val currentCondition: String?,
-    @Json(name = "icon") val weatherIcon: String?,
-)
