+++ app/src/main/java/com/example/android/temptext/network/WeatherResponse.kt
@@ -1,10 +1,23 @@
 package com.example.android.temptext.network
 
 import com.squareup.moshi.Json
-import java.util.concurrent.locks.Condition
 
 class WeatherResponse (
     @Json(name = "location") val currentLocation: NestedWeatherResponse?,
     @Json(name = "current") val currentWeather: NestedWeatherResponse?,
 
+)
+
+class NestedWeatherResponse (
+    @Json(name = "temp_c") val celsius: Float?,
+    @Json(name = "temp_f") val fahrenheit: Float?,
+    @Json(name = "name") val city: String?,
+    @Json(name = "localtime") val localTime: String?,
+    @Json(name = "condition") val currentWeatherCondition: WeatherConditions?,
+    @Json(name = "is_day") val dayOfWeek: Int?
+)
+
+class WeatherConditions (
+    @Json(name = "text") val currentCondition: String?,
+    @Json(name = "icon") val weatherIcon: String?,
 )
